requirements.txt
Django==5.0.2
django_ckeditor==6.7.1
dj-database-url
django_configurations==2.5
django_ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
python-dotenv==1.0.1
django-jazzmin==2.6.0
whitenoise==5.3.0
dj-database-url==0.5.0
gunicorn==20.0.4
pillow==10.2.0
psycopg2-binary
boto3 == 1.16.0
django-storages == 1.14.2
django-cors-headers == 4.3.1
django-extensions == 3.2.3
celery 
redis 
django-celery-beat 
django-celery-results
matplotlib
gevent


